Implement WriteImage, add /_ping endpoint
